Refactored directory structure to more logical one and changed paths in imports etc.
