+++ frontend/src/auxiliary/auth.jsx
