+++ frontend/src/pages/Components/Landing/Home.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 export default function Home() {
     return (
-        <div  className='bg-warning p-3'>
+        <div className='bg-warning p-3'>
             <h1>LANDING ZONE</h1>
             <h2>chcemy:</h2>
             <h3>info co to i po co</h3>
