+++ frontend/src/pages/Components/Register/RegisterHere.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Registration from '../Registration';
+import Registration from './Registration';
 
 export default function RegisterHere() {
     return (
