+++ frontend/src/pages/Components/Register/Registration.js
@@ -3,7 +3,7 @@ import Form from 'react-bootstrap/Form'
 import Button from 'react-bootstrap/Button'
 
 
-import './Registration.css';
+import '../../../styles/Registration.css';
 import RegistrationAlert from "./RegistrationAlert";
 
 class Registration extends Component {
