+++ frontend/src/service/globals/global.js
@@ -1,0 +1,4 @@
+// import React from 'react';
+
+window.$authorized = true;
+// console.log('GOOSFRABAAAAAAAAAAAAAAAAAAA');
