+++ frontend/src/styles/Layout.css
@@ -13,4 +13,5 @@
 .navtitle {
     font-family: "Montserrat", sans-serif;
     font-size: 1.4rem;
+    text-align: left;
 }
