+++ frontend/src/styles/Registration.css
