+++ frontend/src/styles/RegistrationAlert.css
